Fixed player animation, fixed wanted level logic
About the Wanted Level:
The old behaviour was that when you had a wanted level of 0, the script will set it to 2 for you when you started robbing, but if you had more than 2 stars (for example 5), it will overwrite the current wanted level with 2.
Now, if your wanted level is higher than 2, it will keep your current level instead of overwriting it.

About the animation:
The animation requires that the proper animation dictionary is loaded. Now, the script loads the required dict, waits until it's loaded, and then plays the animation.